Separate some responsibilities
Not happy about the on-call import for EMBEDDED_MESSAGE
